Integrate table filters
