+++ package.json
@@ -49,7 +49,7 @@
     "@fundamental-ngx/ui5-webcomponents": "0.59.1",
     "@fundamental-ngx/ui5-webcomponents-fiori": "0.59.1",
     "@openmfp/config-prettier": "0.9.2",
-    "@openmfp/ngx": "file:.yalc/@openmfp/ngx",
+    "@openmfp/ngx": "0.14.2",
     "@openmfp/portal-ui-lib": "0.192.53",
     "@vitest/coverage-v8": "4.1.9",
     "cpx2": "9.0.0",
