--- projects/wc/src/app/components/generic-ui/opensearch-list-view/generic-view/generic-view.component.scss
@@ -1,53 +1,0 @@
-.resource-title {
-  display: flex;
-  align-items: center;
-  padding: 1rem 0;
-}
-
-.resource-title-subheading {
-  display: block;
-  padding-bottom: 1rem;
-}
-
-.resource-logo {
-  width: 5rem;
-  padding-bottom: 1rem;
-}
-
-.resource-title-actions {
-  height: auto;
-}
-
-.resource-info {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 2rem;
-}
-
-.resource-info-cell {
-  display: flex;
-  gap: 5px;
-  flex-direction: column;
-  max-width: 300px;
-}
-
-.multiline {
-  display: block;
-}
-
-.resource-title-snapped-container {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  padding-bottom: 1rem;
-
-  img {
-    width: 50px;
-    height: 50px;
-    padding-bottom: 0;
-  }
-
-  ui5-title {
-    font-size: 1.5rem;
-  }
-}
