+++ projects/wc/src/app/components/generic-ui/opensearch-list-view/open-search-resource-table-card/open-search-resource-table-card.component.html
@@ -1,0 +1,9 @@
+<mfp-declarative-table-card
+  [resources]="resources()"
+  [config]="config()"
+  (tableRowClicked)="navigateToResource($event)"
+  (loadMoreResources)="loadMore()"
+  (paginationLimitChanged)="onLimitChange($event)"
+  (searchChanged)="searchChanged($event)"
+  (filterTabChanged)="onFilterTabChanged($event)"
+/>
